fix refreshing refresh token bug
