+++ src/auth/routers/auth.router.ts
@@ -2,12 +2,13 @@ import { Router } from "express";
 import { validateRequestProperty } from "../../validation/middleware/validateRequestProperty";
 import { UserSchema } from "../../validation/schemas/user.schema";
 import signUp from "../middleware/signUp";
-import sendAuthenticationResponse from "../middleware/sendAuthenticationResponse";
+import sendTokens from "../middleware/sendTokens";
+import refreshAccessToken from "../middleware/refreshAccessToken";
 
 const authRouter = Router();
 
 authRouter.post("/signup", validateRequestProperty("body", UserSchema), signUp);
-authRouter.post("/login", sendAuthenticationResponse("local"));
-authRouter.post("/refresh", sendAuthenticationResponse("jwt-refresh"));
+authRouter.post("/login", sendTokens);
+authRouter.post("/refresh", refreshAccessToken);
 
 export default authRouter;
